requirements.txt
attrs==25.3.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
mdformat==0.7.22
mdformat-tables==1.0.0
mdit_py_plugins==0.4.2
python-frontmatter==1.1.0
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.25.1
